+++ packages/expo-audio/src/index.ts
@@ -40,23 +40,18 @@ export function useAudioSampleListener(
   player.setAudioSamplingEnabled(true);
   useEffect(() => {
     const subscription = player.addListener('onAudioSampleUpdate', listener);
-    return () => {
-      player.setAudioSamplingEnabled(false);
-      subscription.remove();
-    };
+    return () => subscription.remove();
   }, [player.id]);
 }
 
 export function useAudioRecorder(
   options: RecordingOptions,
   statusListener?: (status: RecordingStatus) => void
-): [AudioRecorder, RecorderState] {
+): AudioRecorder {
   const platformOptions = createRecordingOptions(options);
   const recorder = useReleasingSharedObject(() => {
     return new AudioModule.AudioRecorder(platformOptions);
   }, [JSON.stringify(platformOptions)]);
-
-  const [state, setState] = useState<RecorderState>(recorder.getStatus());
 
   useEffect(() => {
     const subscription = recorder.addListener('onRecordingStatusUpdate', (status) => {
@@ -65,16 +60,21 @@ export function useAudioRecorder(
     return () => subscription.remove();
   }, [recorder.id]);
 
-  useEffect(() => {
-    const interval = setInterval(() => {
-      const status = recorder.getStatus();
-      setState(status);
-    }, 1000);
+  return recorder;
+}
 
-    return () => clearInterval(interval);
+export function useAudioRecorderState(recorder: AudioRecorder, interval: number = 1000) {
+  const [state, setState] = useState<RecorderState>(recorder.getStatus());
+
+  useEffect(() => {
+    const int = setInterval(() => {
+      setState(recorder.getStatus());
+    }, interval);
+
+    return () => clearInterval(int);
   }, [recorder.id]);
 
-  return [recorder, state];
+  return state;
 }
 
 export async function setIsAudioActiveAsync(active: boolean): Promise<void> {