+++ packages/expo-modules-core/android/src/main/cpp/JavaReferencesCache.cpp
@@ -37,7 +37,8 @@ void JavaReferencesCache::loadJClasses(JNIEnv *env) {
   loadJClass(env, "[J", {});
   loadJClass(env, "[F", {});
 
-  loadJClass(env, "java/util/List", {});
+  loadJClass(env, "java/util/Collection", {});
+  loadJClass(env, "java/util/Map", {});
 
   loadJClass(env, "com/facebook/react/bridge/PromiseImpl", {
     {"<init>", "(Lcom/facebook/react/bridge/Callback;Lcom/facebook/react/bridge/Callback;)V"}